restructuring source files, adapting to new mpc with disturbance
